+++ src/datasources/results/components/editors/query-steps/QueryStepsEditor.test.tsx
@@ -290,6 +290,18 @@ describe('QueryStepsEditor', () => {
       );
     });
 
+    test('should not update results query when filter doesnt change', () => {
+      const resultsQueryInput = screen.getByTestId('results-query');
+      fireEvent.change(resultsQueryInput, { target: { value: 'partNumber = "PN1"' } });// ensure initial value is set
+      mockHandleQueryChange.mockClear();
+
+      fireEvent.change(resultsQueryInput, { target: { value: 'partNumber = "PN1"' } });
+
+      expect(mockHandleQueryChange).not.toHaveBeenCalledWith(
+        expect.objectContaining({ resultsQuery: 'partNumber = "PN1"' })
+      );
+    });
+
     test('should update steps query when user triggers steps query change', () => {
       const stepsQueryInput = screen.getByTestId('steps-query');
 
@@ -299,6 +311,18 @@ describe('QueryStepsEditor', () => {
         expect.objectContaining({ stepsQuery: 'updated-steps-query' })
       );
     });
+
+    test('should not update steps query when filter doesnt change', () => {
+      const stepsQueryInput = screen.getByTestId('steps-query');
+      fireEvent.change(stepsQueryInput, { target: { value: 'stepName = "Step1"' } });// ensure initial value is set
+      mockHandleQueryChange.mockClear();
+
+      fireEvent.change(stepsQueryInput, { target: { value: 'stepName = "Step1"' } });
+
+      expect(mockHandleQueryChange).not.toHaveBeenCalledWith(
+        expect.objectContaining({ stepsQuery: 'stepName = "Step1"' })
+      );
+    })
 
     test('should disable steps query builder when partnumber is empty', async () => {
       cleanup();