+++ src/datasources/results/query-handlers/query-steps/QueryStepsDataSource.ts
@@ -43,7 +43,7 @@ export class QueryStepsDataSource extends ResultsDataSourceBase {
     projection?: StepsProperties[],
     take?: number,
     descending?: boolean,
-    resultsFilter?: string,
+    resultFilter?: string,
     continuationToken?: string,
     returnCount = false
   ): Promise<QueryStepsResponse> {
@@ -55,7 +55,7 @@ export class QueryStepsDataSource extends ResultsDataSourceBase {
         descending,
         projection,
         take,
-        resultsFilter,
+        resultFilter,
         continuationToken,
         returnCount,
       });
@@ -104,7 +104,7 @@ export class QueryStepsDataSource extends ResultsDataSourceBase {
     projection?: StepsProperties[],
     take?: number,
     descending?: boolean,
-    resultsFilter?: string,
+    resultFilter?: string,
     returnCount = false,
   ): Promise<QueryStepsResponse> {
     const queryRecord = async (currentTake: number, token?: string): Promise<QueryResponse<StepsResponseProperties>> => {
@@ -114,7 +114,7 @@ export class QueryStepsDataSource extends ResultsDataSourceBase {
         projection,
         currentTake,
         descending,
-        resultsFilter,
+        resultFilter,
         token,
         returnCount
       );