+++ src/datasources/test-plans/TestPlansDataSource.ts
@@ -10,7 +10,7 @@ import { AssetUtils } from './asset.utils';
 import { WorkspaceUtils } from 'shared/workspace.utils';
 import { SystemUtils } from 'shared/system.utils';
 import { QueryBuilderOperations } from 'core/query-builder.constants';
-import { ExpressionTransformFunction, transformComputedFieldsQuery } from 'core/query-builder.utils';
+import { computedFieldsupportedOperations, ExpressionTransformFunction, transformComputedFieldsQuery } from 'core/query-builder.utils';
 import { UsersUtils } from 'shared/users.utils';
 import { ProductUtils } from 'shared/product.utils';
 
@@ -68,6 +68,7 @@ export class TestPlansDataSource extends DataSourceBase<TestPlansQuery> {
         this.templateSrv.replace(query.queryBy, options.scopedVars),
         this.testPlansComputedDataFields
       );
+      query.queryBy = this.transformDurationFilters(query.queryBy);
     }
 
     if (query.outputType === OutputType.Properties) {
@@ -183,6 +184,21 @@ export class TestPlansDataSource extends DataSourceBase<TestPlansQuery> {
     return true;
   }
 
+  private transformDurationFilters(query: string): string {
+    const daysRegex = new RegExp(`estimatedDurationInDays\\s*(${computedFieldsupportedOperations.join('|')})\\s*"(\\d+)"`, 'g');
+    const hoursRegex = new RegExp(`estimatedDurationInHours\\s*(${computedFieldsupportedOperations.join('|')})\\s*"(\\d+)"`, 'g');
+
+    return query
+      .replace(
+        daysRegex,
+        (_, operator, value) => `estimatedDurationInSeconds ${operator} "${parseInt(value, 10) * 86400}"`
+      )
+      .replace(
+        hoursRegex,
+        (_, operator, value) => `estimatedDurationInSeconds ${operator} "${parseInt(value, 10) * 3600}"`
+      );
+  }
+
   private async getFixtureNames(labels: string[], testPlans: TestPlanResponseProperties[]): Promise<Asset[]> {
     if (labels.find(label => label === PropertiesProjectionMap.FIXTURE_NAMES.label)) {
       const fixtureIds = testPlans
@@ -230,12 +246,14 @@ export class TestPlansDataSource extends DataSourceBase<TestPlansQuery> {
   }
 
   async metricFindQuery(query: TestPlansVariableQuery, options: LegacyMetricFindQueryOptions): Promise<MetricFindValue[]> {
-    const filter = query.queryBy ?
-      transformComputedFieldsQuery(
+    let filter;
+    if (query.queryBy) {
+      filter = transformComputedFieldsQuery(
         this.templateSrv.replace(query.queryBy, options.scopedVars),
         this.testPlansComputedDataFields
-      )
-      : undefined;
+      );
+      filter = this.transformDurationFilters(filter);
+    }
 
     const metadata = (await this.queryTestPlansInBatches(
       filter,