+++ src/datasources/work-orders/components/query-builder/WorkOrdersQueryBuilder.tsx
@@ -1,19 +1,25 @@
 import { SlQueryBuilder } from 'core/components/SlQueryBuilder/SlQueryBuilder';
 import { queryBuilderMessages, QueryBuilderOperations } from 'core/query-builder.constants';
 import { expressionBuilderCallback, expressionReaderCallback } from 'core/query-builder.utils';
-import { QBField, QueryBuilderOption } from 'core/types';
+import { QBField, QueryBuilderOption, Workspace } from 'core/types';
 import { filterXSSField } from 'core/utils';
 import { WorkOrdersQueryBuilderFields, WorkOrdersQueryBuilderStaticFields } from 'datasources/work-orders/constants/WorkOrdersQueryBuilder.constants';
 import React, { useState, useEffect, useMemo } from 'react';
+import { User } from 'shared/types/QueryUsers.types';
+import { UsersUtils } from 'shared/users.utils';
 import { QueryBuilderCustomOperation, QueryBuilderProps } from 'smart-webcomponents-react/querybuilder';
 
 type WorkOrdersQueryBuilderProps = QueryBuilderProps & React.HTMLAttributes<Element> & {
   filter?: string;
+  workspaces: Workspace[] | null;
+  users: User[] | null;
   globalVariableOptions: QueryBuilderOption[];
 };
 
 export const WorkOrdersQueryBuilder: React.FC<WorkOrdersQueryBuilderProps> = ({
   filter,
+  workspaces,
+  users,
   onChange,
   globalVariableOptions,
 }) => {
@@ -49,8 +55,35 @@ export const WorkOrdersQueryBuilder: React.FC<WorkOrdersQueryBuilderProps> = ({
     ];
   }, []);
 
+  const workspaceField = useMemo(() => {
+    if (!workspaces) {
+        return null;
+    }
+    const workspaceOptions = workspaces.map(({ id, name }) => ({ label: name, value: id }));
+
+    return addOptionsToLookup(WorkOrdersQueryBuilderFields.WORKSPACE, workspaceOptions);
+  }, [workspaces]);
+
+  const usersField = useMemo(() => {
+    if (!users) {
+      return;
+    }    
+    const usersMap = users.map(user => ({ label: UsersUtils.getUserNameAndEmail(user), value: user.id }));
+
+    return [
+      addOptionsToLookup(WorkOrdersQueryBuilderFields.ASSIGNED_TO, usersMap),
+      addOptionsToLookup(WorkOrdersQueryBuilderFields.CREATED_BY, usersMap),
+      addOptionsToLookup(WorkOrdersQueryBuilderFields.REQUESTED_BY, usersMap),
+      addOptionsToLookup(WorkOrdersQueryBuilderFields.UPDATED_BY, usersMap),
+    ];
+  }, [users]);
+
   useEffect(() => {
-    const updatedFields = [...WorkOrdersQueryBuilderStaticFields, ...timeFields].map(field => {
+    if (!workspaceField || !timeFields || !usersField) {
+      return;
+    }
+
+    const updatedFields = [...WorkOrdersQueryBuilderStaticFields, ...timeFields, workspaceField, ...usersField].map(field => {
       if (field.lookup?.dataSource) {
         return {
           ...field,
@@ -105,7 +138,7 @@ export const WorkOrdersQueryBuilder: React.FC<WorkOrdersQueryBuilderProps> = ({
     ];
 
     setOperations([...customOperations, ...keyValueOperations]);
-  }, [globalVariableOptions, timeFields]);
+  }, [globalVariableOptions, timeFields, workspaceField, usersField]);
 
   return (
     <SlQueryBuilder