+++ CHANGELOG.md
@@ -1,3 +1,53 @@
+## [3.125.1](https://github.com/ni/systemlink-grafana-plugins/compare/v3.125.0...v3.125.1) (2025-06-10)
+
+### Bug Fixes
+
+* **results:** `totalCount` is not returned when there are no results matching the query ([#266](https://github.com/ni/systemlink-grafana-plugins/issues/266)) ([f7095ec](https://github.com/ni/systemlink-grafana-plugins/commit/f7095ec51091ebc01e4dc1c51cf95b49216afce2))
+
+## [3.125.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.124.1...v3.125.0) (2025-06-10)
+
+### Features
+
+* **results:** rename resultsFilter to resultFilter ([#264](https://github.com/ni/systemlink-grafana-plugins/issues/264)) ([782025f](https://github.com/ni/systemlink-grafana-plugins/commit/782025f8c0cf2dc36a50fa8e85f63f532f3bc5f8))
+* **test-plans:** Add state property lookup in the query builder ([#263](https://github.com/ni/systemlink-grafana-plugins/issues/263)) ([c7dacd8](https://github.com/ni/systemlink-grafana-plugins/commit/c7dacd822159775a91f4517e1750cf230d364789))
+
+## [3.124.1](https://github.com/ni/systemlink-grafana-plugins/compare/v3.124.0...v3.124.1) (2025-06-10)
+
+### Bug Fixes
+
+* **testplans:** Update filter fields in test plans ([#260](https://github.com/ni/systemlink-grafana-plugins/issues/260)) ([d73b005](https://github.com/ni/systemlink-grafana-plugins/commit/d73b005e5c774e75152309d21de13f0530cb4a39))
+
+## [3.124.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.123.0...v3.124.0) (2025-06-10)
+
+### Features
+
+* **workorders:** Add users lookup and map users in workorders data ([#256](https://github.com/ni/systemlink-grafana-plugins/issues/256)) ([1d74091](https://github.com/ni/systemlink-grafana-plugins/commit/1d74091fb3b7d9631d13a2529a8679fba94d8fdf))
+
+## [3.123.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.122.0...v3.123.0) (2025-06-10)
+
+### Features
+
+* **workorders:** Map workspace name in workorders QB and datasource ([#255](https://github.com/ni/systemlink-grafana-plugins/issues/255)) ([f76916b](https://github.com/ni/systemlink-grafana-plugins/commit/f76916b9647adf19eecb7426b1d2a194a92bef57))
+
+## [3.122.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.121.0...v3.122.0) (2025-06-10)
+
+### Features
+
+* **results:** Handle null values for product names ([#262](https://github.com/ni/systemlink-grafana-plugins/issues/262)) ([a397fc3](https://github.com/ni/systemlink-grafana-plugins/commit/a397fc37e115b2dfead0ffe12a586425921784df))
+
+## [3.121.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.120.0...v3.121.0) (2025-06-10)
+
+### Features
+
+* **results:** prevent duplicate API call  ([#259](https://github.com/ni/systemlink-grafana-plugins/issues/259)) ([373da70](https://github.com/ni/systemlink-grafana-plugins/commit/373da70efa07aa8bc0e903b27e6b51bfbc1c62b1))
+
+## [3.120.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.119.0...v3.120.0) (2025-06-10)
+
+### Features
+
+* **test-plans:** Add users lookup and map users in test plans data source & query builder ([#257](https://github.com/ni/systemlink-grafana-plugins/issues/257)) ([4eb5e3e](https://github.com/ni/systemlink-grafana-plugins/commit/4eb5e3e8756e5a7efaba1b2f813ffc0077f91033))
+* **test-plans:** Show Dut serial number in the properties ([#258](https://github.com/ni/systemlink-grafana-plugins/issues/258)) ([8934fe5](https://github.com/ni/systemlink-grafana-plugins/commit/8934fe5faec9350d293e787745c8de92c2410bb0))
+
 ## [3.119.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.118.0...v3.119.0) (2025-06-09)
 
 ### Features